PHP Data Objects (PDO)
PDO is an interface for accessing databases using PHP. It allows you to
access a variety of different databases depending on which driver is
installed, and provides an easy way to query, retrieve results, prepare
SQL statements and bake database security into your application. In this
series we look at the basics of PDO to get you started.